+++ lib/include/mb_mgr_job_api.h
@@ -58,6 +58,9 @@
                 func(state, job->src + job->hash_start_src_offset_in_bytes,                        \
                      job->msg_len_to_hash_in_bytes)
 
+static const uint8_t zero_low_4B_mask[16] = { 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
+                                              0xFF, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00 };
+
 /* ========================================================================= */
 /* AES-CBC */
 /* ========================================================================= */
@@ -203,6 +206,25 @@ SUBMIT_JOB_SM4_GCM_ENC(IMB_MGR *state, IMB_JOB *job)
                 job->auth_tag_output_len_in_bytes, IMB_DIR_ENCRYPT);
 
         job->status |= IMB_STATUS_COMPLETED;
+
+        return job;
+}
+
+/* ========================================================================= */
+/* AES-NEA5 */
+/* ========================================================================= */
+__forceinline IMB_JOB *
+SUBMIT_JOB_AES_NEA5(IMB_JOB *job)
+{
+        uint8_t iv[16];
+
+        __m128i iv_reg = _mm_loadu_si128((const __m128i *) job->iv);
+        iv_reg = _mm_and_si128(iv_reg, _mm_loadu_si128((const __m128i *) zero_low_4B_mask));
+
+        _mm_storeu_si128((__m128i *) iv, iv_reg);
+        AES_CTR_256(job->src + job->cipher_start_src_offset_in_bytes, iv, job->enc_keys, job->dst,
+                    job->msg_len_to_cipher_in_bytes, 16);
+        job->status |= IMB_STATUS_COMPLETED_CIPHER;
 
         return job;
 }
@@ -405,6 +427,8 @@ SUBMIT_JOB_CIPHER_ENC(IMB_MGR *state, IMB_JOB *job, const IMB_CIPHER_MODE cipher
                 }
         } else if (IMB_CIPHER_CNTR == cipher_mode) {
                 return SUBMIT_JOB_AES_CTR(job, key_sz);
+        } else if (IMB_CIPHER_AES_NEA5 == cipher_mode) {
+                return SUBMIT_JOB_AES_NEA5(job);
         } else if (IMB_CIPHER_ECB == cipher_mode) {
                 if (16 == key_sz) {
                         return SUBMIT_JOB_AES_ECB_128_ENC(job);
@@ -625,6 +649,8 @@ SUBMIT_JOB_CIPHER_DEC(IMB_MGR *state, IMB_JOB *job, const IMB_CIPHER_MODE cipher
                 }
         } else if (IMB_CIPHER_CNTR == cipher_mode) {
                 return SUBMIT_JOB_AES_CTR(job, key_sz);
+        } else if (IMB_CIPHER_AES_NEA5 == cipher_mode) {
+                return SUBMIT_JOB_AES_NEA5(job);
         } else if (IMB_CIPHER_ECB == cipher_mode) {
                 if (16 == key_sz) {
                         return SUBMIT_JOB_AES_ECB_128_DEC(job);
@@ -1027,6 +1053,13 @@ submit_cipher_dec_cfb_256(IMB_MGR *state, IMB_JOB *job)
         return SUBMIT_JOB_CIPHER_DEC(state, job, IMB_CIPHER_CFB, IMB_KEY_256_BYTES);
 }
 
+/* AES-NEA5 */
+static IMB_JOB *
+submit_cipher_dec_aes_nea5(IMB_MGR *state, IMB_JOB *job)
+{
+        return SUBMIT_JOB_CIPHER_DEC(state, job, IMB_CIPHER_AES_NEA5, IMB_KEY_256_BYTES);
+}
+
 /* ========================= */
 /* ======== ENCRYPT ======== */
 /* ========================= */
@@ -1284,6 +1317,14 @@ submit_cipher_enc_cfb_256(IMB_MGR *state, IMB_JOB *job)
 {
         return SUBMIT_JOB_CIPHER_ENC(state, job, IMB_CIPHER_CFB, IMB_KEY_256_BYTES);
 }
+
+/* AES-NEA5 */
+static IMB_JOB *
+submit_cipher_enc_aes_nea5(IMB_MGR *state, IMB_JOB *job)
+{
+        return SUBMIT_JOB_CIPHER_ENC(state, job, IMB_CIPHER_AES_NEA5, IMB_KEY_256_BYTES);
+}
+
 /*
  * Four entries per algorithm (different key sizes),
  * algorithms in the same order IMB_CIPHER_MODE
@@ -1436,12 +1477,12 @@ static const submit_flush_fn_t tab_submit_cipher[] = {
         submit_cipher_dec_snow5g,
         submit_cipher_dec_snow5g,
         submit_cipher_dec_snow5g,
+        /* [27] AES NEA5 */
+        submit_cipher_dec_null,
+        submit_cipher_dec_null,
+        submit_cipher_dec_null,
+        submit_cipher_dec_aes_nea5,
         /* add new cipher decrypt here */
-        /* [27] NULL */
-        NULL,
-        NULL,
-        NULL,
-        NULL,
         /* [28] NULL */
         NULL,
         NULL,
@@ -1602,12 +1643,12 @@ static const submit_flush_fn_t tab_submit_cipher[] = {
         submit_cipher_enc_snow5g,
         submit_cipher_enc_snow5g,
         submit_cipher_enc_snow5g,
+        /* [27] AES NEA5 */
+        submit_cipher_enc_null,
+        submit_cipher_enc_null,
+        submit_cipher_enc_null,
+        submit_cipher_enc_aes_nea5,
         /* add new cipher encrypt here */
-        /* [27] NULL */
-        NULL,
-        NULL,
-        NULL,
-        NULL,
         /* [28] NULL */
         NULL,
         NULL,
@@ -1903,6 +1944,14 @@ flush_cipher_dec_cfb_256(IMB_MGR *state, IMB_JOB *job)
 {
         return FLUSH_JOB_CIPHER_DEC(state, job, IMB_CIPHER_CFB, IMB_KEY_256_BYTES);
 }
+
+/* AES-NEA5 */
+static IMB_JOB *
+flush_cipher_dec_aes_nea5(IMB_MGR *state, IMB_JOB *job)
+{
+        return FLUSH_JOB_CIPHER_DEC(state, job, IMB_CIPHER_AES_NEA5, IMB_KEY_256_BYTES);
+}
+
 /* ========================= */
 /* ======== ENCRYPT ======== */
 /* ========================= */
@@ -2155,7 +2204,7 @@ flush_cipher_enc_sm4_gcm(IMB_MGR *state, IMB_JOB *job)
         return FLUSH_JOB_CIPHER_ENC(state, job, IMB_CIPHER_SM4_GCM, IMB_KEY_128_BYTES);
 }
 
-/* AES-CBC */
+/* AES-CFB */
 static IMB_JOB *
 flush_cipher_enc_cfb_128(IMB_MGR *state, IMB_JOB *job)
 {
@@ -2171,6 +2220,14 @@ flush_cipher_enc_cfb_256(IMB_MGR *state, IMB_JOB *job)
 {
         return FLUSH_JOB_CIPHER_ENC(state, job, IMB_CIPHER_CFB, IMB_KEY_256_BYTES);
 }
+
+/* AES-NEA5 */
+static IMB_JOB *
+flush_cipher_enc_aes_nea5(IMB_MGR *state, IMB_JOB *job)
+{
+        return FLUSH_JOB_CIPHER_ENC(state, job, IMB_CIPHER_AES_NEA5, IMB_KEY_256_BYTES);
+}
+
 /*
  * Four entries per algorithm (different key sizes),
  * algorithms in the same order IMB_CIPHER_MODE
@@ -2318,12 +2375,12 @@ static const submit_flush_fn_t tab_flush_cipher[] = {
         flush_cipher_dec_snow5g,
         flush_cipher_dec_snow5g,
         flush_cipher_dec_snow5g,
+        /* [27] AES NEA5 */
+        flush_cipher_dec_null,
+        flush_cipher_dec_null,
+        flush_cipher_dec_null,
+        flush_cipher_dec_aes_nea5,
         /* add new cipher decrypt here */
-        /* [27] NULL */
-        NULL,
-        NULL,
-        NULL,
-        NULL,
         /* [28] NULL */
         NULL,
         NULL,
@@ -2484,12 +2541,12 @@ static const submit_flush_fn_t tab_flush_cipher[] = {
         flush_cipher_enc_snow5g,
         flush_cipher_enc_snow5g,
         flush_cipher_enc_snow5g,
+        /* [27] AES NEA5 */
+        flush_cipher_enc_null,
+        flush_cipher_enc_null,
+        flush_cipher_enc_null,
+        flush_cipher_enc_aes_nea5,
         /* add new cipher encrypt here */
-        /* [27] NULL */
-        NULL,
-        NULL,
-        NULL,
-        NULL,
         /* [28] NULL */
         NULL,
         NULL,