+++ ReleaseNotes.md
@@ -25,6 +25,7 @@ Library
 - Added ZUC-NEA6 SSE, AVX2 and AVX512 implementations.
 - Added SNOW5G-NEA4 SSE type 1 implementation.
 - Added AES-NIA5 SSE type 1 implementation.
+- Added AES-NEA5 SSE, AVX2 and AVX512 implementations.
 
 Test Applications
 - Added `--offset` parameter to `imb-xvalid` application.